Optimizing AdbClient's use of sock.recv
The __receive, __readExactly, and shell methods of AdbClient all used
inefficient strategies for receiving large buffers. For the reasons
explained below, all three would perform at O(n^2). While __receive
was much more performant than __readExactly and shell in practice,
the worst-case performance was still O(n^2).

The __readExactly, and shell functions both used string concatenation
when using buffers larger that 4k. Because strings are immutable in
Python, this results in a new allocation and a copy of the previous
buffer for each concatenation. The results is O(n^2) performance.

The __receive method used a bytearray with extend. Since bytearray
objects use a realloc to resize the buffer when calling extend, and
may allocate more space than required for the extend, calling extend
doesn't necessarily mean that realloc will be called, and a realloc
doesn't necessarily mean a new allocation and a copy of all old data.
This is, however, highly dependent on the state of the heap in
the Python process. In cases where realloc in called every time, and
a new allocation is required, the performance is O(n^2). In cases
where realloc allocates more than the required space, this may
approach O(n*log(n)), but is implementation specific.

For all of these scenarios we can move performance to O(n). In the
cases that we know the buffer size, We can reduce the number of
allocations from n to 1, by allocating the size we need upfront. We
can then reduce data copies to O(n) by reading directly into the
buffer. In cases where the size is unknown, we must keep the number
of allocations at O(n), but we can reduce the data copies to O(n)
from O(n^2) by using a join operation.